requirements.txt
GoogleNews==1.6.8
matplotlib==3.7.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
scikit_learn==1.3.0
scipy==1.10.0
seaborn==0.12.2
streamlit==1.22.0
yfinance==0.2.23
mpld3
jugaad_data